armv8a: mmu: Make driver more CMSIS-friendly
