+++ mcux/mcux-sdk/CMSIS/Core_AArch64/Include/CMSIS_Include_core_ca.cmake
@@ -3,7 +3,7 @@ include_guard(GLOBAL)
 message("CMSIS_Include_core_ca component is included.")
 
 target_sources(${MCUX_SDK_PROJECT_NAME} PRIVATE
-    ${CMAKE_CURRENT_LIST_DIR}/../Source/arm_mmu.c
+    ${CMAKE_CURRENT_LIST_DIR}/../Source/mmu_armv8a.c
 )
 
 target_include_directories(${MCUX_SDK_PROJECT_NAME} PUBLIC
