+++ mcux/mcux-sdk/CMSIS/Core_AArch64/Include/core_ca53.h
@@ -126,17 +126,6 @@
 
 
 /*******************************************************************************
- *                 MMU Functions
- ******************************************************************************/
-
-#if defined (__MMU_PRESENT) && (__MMU_PRESENT == 1U)
-
-  #include "arm_mmu.h"
-
-#endif
-
-
-/*******************************************************************************
  *                 GIC Functions
  ******************************************************************************/
 
@@ -148,11 +137,20 @@
 
 
 /*******************************************************************************
+ *                 MMU Functions
+ ******************************************************************************/
+
+#if defined (__MMU_PRESENT) && (__MMU_PRESENT == 1U)
+
+  #include "mmu_armv8a.h"
+
+#endif
+
+
+/*******************************************************************************
  *                 Timer Functions
  ******************************************************************************/
 
-#if defined (__TIM_PRESENT)
-#endif
 #if defined (__TIM_PRESENT) && (__TIM_PRESENT == 1U)
   #include "timer_armv8a.h"
 #endif
